API追加、API名称変更、JOOQのログ出力設定

+++ infra/docker/db/sql/01_initial_db.sql
@@ -1,6 +1,6 @@
 CREATE DATABASE IF NOT EXISTS tsk2 DEFAULT CHARACTER SET utf8;
 CREATE TABLE IF NOT EXISTS tsk2.tasks(
-  id INT PRIMARY KEY AUTO_INCREMENT,
+  id BIGINT PRIMARY KEY AUTO_INCREMENT,
   title VARCHAR(100),
   created_at TIMESTAMP NOT NULL default current_timestamp,
   updated_at TIMESTAMP NOT NULL default current_timestamp on update current_timestamp
